+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsactivitiesmanagementapi/entity/EventConsumed.kt
@@ -12,7 +12,7 @@ import javax.persistence.Table
 data class EventConsumed(
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  val eventId: Int? = null,
+  val eventId: Long? = null,
 
   val eventType: String,
